IS: Refactor routing for sykmeldinger page
